Move composables into their own files, make some minor refactoring changes.
