+++ app/src/main/java/android/bignerdranch/codapizza/ui/PizzaBuilderScreen.kt
@@ -1,29 +1,21 @@
 package android.bignerdranch.codapizza.ui
 
-import android.bignerdranch.codapizza.R
 import android.bignerdranch.codapizza.model.Pizza
-import android.bignerdranch.codapizza.model.Topping
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import java.text.NumberFormat
 
 @Preview
 @Composable
@@ -34,16 +26,7 @@ fun PizzaBuilderScreen(
     Column(
         modifier = modifier
     ){
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(color = MaterialTheme.colorScheme.secondaryContainer)
-                .padding(16.dp)
-        ){
-            Text(
-                text = "Coda Pizza by Greg Murashige"
-            )
-        }
+        Header()
         PizzaSizeDropdown(
             pizza = pizza,
             onEditPizza = { pizza = it },
@@ -66,56 +49,15 @@ fun PizzaBuilderScreen(
 }
 
 @Composable
-private fun ToppingsList(
-    pizza: Pizza,
-    onEditPizza: (Pizza) -> Unit,
-    modifier: Modifier = Modifier
-){
-    var toppingBeingAdded by rememberSaveable { mutableStateOf<Topping?>(null) }
-    toppingBeingAdded?.let { topping ->
-        ToppingPlacementDialog(
-            topping = topping,
-            onSetToppingPlacement = { placement ->
-                onEditPizza(pizza.withTopping(topping, placement))
-            },
-            onDismissRequest = {
-                toppingBeingAdded = null
-            }
-        )
-    }
-    LazyColumn(
-        modifier = modifier
-    ) {
-        items(Topping.entries.toTypedArray()) { topping ->
-            ToppingCell(
-                topping = topping,
-                placement = pizza.toppings[topping],
-                onClickTopping = {
-                    toppingBeingAdded = topping
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun OrderButton(
-    pizza: Pizza,
-    modifier: Modifier = Modifier
-){
-    Button(
-        modifier = modifier,
-        onClick = {
-            // TODO
-        }
-    ) {
-        val currencyFormatter = remember { NumberFormat.getCurrencyInstance() }
-        val price = currencyFormatter.format(pizza.price)
+private fun Header(){
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(color = MaterialTheme.colorScheme.secondaryContainer)
+            .padding(16.dp)
+    ){
         Text(
-            text = stringResource(R.string.place_order_button,
-                price,
-                stringResource(pizza.size.sizeName))
-                .uppercase()
+            text = "Coda Pizza by Greg Murashige"
         )
     }
 }
