+++ app/src/main/java/android/bignerdranch/codapizza/ui/ToppingList.kt
@@ -1,0 +1,45 @@
+package android.bignerdranch.codapizza.ui
+
+import android.bignerdranch.codapizza.model.Pizza
+import android.bignerdranch.codapizza.model.Topping
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ToppingsList(
+    pizza: Pizza,
+    onEditPizza: (Pizza) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    var toppingBeingAdded by rememberSaveable { mutableStateOf<Topping?>(null) }
+    toppingBeingAdded?.let { topping ->
+        ToppingPlacementDialog(
+            topping = topping,
+            onSetToppingPlacement = { placement ->
+                onEditPizza(pizza.withTopping(topping, placement))
+            },
+            onDismissRequest = {
+                toppingBeingAdded = null
+            }
+        )
+    }
+    LazyColumn(
+        modifier = modifier
+    ) {
+        items(Topping.entries.toTypedArray()) { topping ->
+            ToppingCell(
+                topping = topping,
+                placement = pizza.toppings[topping],
+                onClickTopping = {
+                    toppingBeingAdded = topping
+                }
+            )
+        }
+    }
+}
